Ignore pep8 F821 fired by guarded Python3 tests executed in Python2
First, I do not like the idea of masking Python3 capabilities (exceptions in this case)
just because we support Python2, too. Second, I want to test it properly, meaning separate
tests for Python2 and Python3 where the details differ.

Perhaps there is a cleaner way to test version-specific behavior and pass all pep8 tests,
but I do not know it.