Make index_len argument to idata optional
PR #92 added an extra argument to the `idata` decorator.  The new
`index_len` argument had no default value, and so the calling convention
was incompatible with older releases of `ddt`.  This commit makes the
second argument optional, with the previous default value inferred if it
is not supplied.  The single-argument form of `idata` will now work
again, while the newer method allowing it to be overridden is still
supported.